requirements.txt
python-binance==1.0.16
python-dotenv==0.19.2
pandas==1.3.5
mplfinance==0.12.9b0
matplotlib==3.5.1
numpy==1.21.0
python-telegram-bot==20.3